force push added SelectionByDisplacement

+++ Code/Facade/GeneticAlgorithm/AbstractRecombinator/AbstractRecombinator.py
@@ -10,10 +10,10 @@ class AbstractRecombinator(Protocol):
 
     _parents_pairs: list  # Список пар родителей
     _selected_children: list  # Список выбранных потомков
-    _crossing_probability: int  # Вероятность рекомбинации
-    _changing_probability: int  # Вероятность изменения гена
+    _crossing_probability: float  # Вероятность рекомбинации
+    _changing_probability: float  # Вероятность изменения гена
 
-    def __init__(self, parents_pairs: list, crossing_probability: int, changing_probability: int):
+    def __init__(self, parents_pairs: list, crossing_probability: float, changing_probability: float):
         '''
         Инициализация экземпляра класса AbstractRecombinator.
